Global exceptions file (#1151)
* initial exceptions.py

* Rename SolverException

* smtlib error

* smt error refactoring

* misc

* SolverError properly inherit

* revert solvererror rename to make pr simpler
