+++ manticore/core/executor.py
@@ -4,10 +4,10 @@ import random
 import logging
 import signal
 
-from ..manticore import ManticoreError
+from ..exceptions import ExecutorError, SolverException
 from ..utils.nointerrupt import WithKeyboardInterruptAs
 from ..utils.event import Eventful
-from .smtlib import Z3Solver, Expression, SolverException
+from .smtlib import Z3Solver, Expression
 from .state import Concretize, TerminateState
 
 from .workspace import Workspace
@@ -15,10 +15,6 @@ from multiprocessing.managers import SyncManager
 from contextlib import contextmanager
 
 # This is the single global manager that will handle all shared memory among workers
-
-
-class ExecutorError(ManticoreError):
-    pass
 
 
 def mgr_init():
